requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
click==8.2.1
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
idna==3.10
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
multidict==6.6.4
platformdirs==4.3.8
propcache==0.3.2
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyyaml==6.0.2
rich==14.1.0
sniffio==1.3.1
starlette==0.47.2
textual==5.3.0
typing-extensions==4.14.1
typing-inspection==0.4.1
uc-micro-py==1.0.3
uvicorn==0.35.0
yarl==1.20.1
